added password encryption
